+++ packages/sdk/src/sdk/merkleProver/local.ts
@@ -1,26 +1,15 @@
 import { Wallet, Wallet__factory } from "@nocturne-xyz/contracts";
-import {
-  InsertNoteCommitmentsEvent,
-  InsertNotesEvent,
-} from "@nocturne-xyz/contracts/dist/src/Wallet";
 import { ethers } from "ethers";
 import { Address } from "../../commonTypes";
 import { BinaryPoseidonTree } from "../../primitives/binaryPoseidonTree";
 import { LocalMerkleDBExtension } from "../db";
-import { query } from "../utils";
 import { MerkleProver } from ".";
 import { MerkleProof } from "@zk-kit/incremental-merkle-tree";
 import { NoteTrait } from "../note";
+import { fetchInsertions } from "../../indexing";
 
 const DEFAULT_START_BLOCK = 0;
 const MERKLE_NEXT_BLOCK_TO_INDEX = "MERKLE_NEXT_BLOCK_TO_INDEX";
-
-interface OrderedLeaf {
-  leaf: bigint;
-  blockNumber: number;
-  txIdx: number;
-  logIdx: number;
-}
 
 export class LocalMerkleProver extends MerkleProver {
   readonly localTree: BinaryPoseidonTree;
@@ -97,76 +86,15 @@ export class LocalMerkleProver extends MerkleProver {
   }
 
   async fetchNewLeaves(from: number, to: number): Promise<bigint[]> {
-    // fetch both kind of insertion events (note commitments and full notes)
-    const ncEventsProm: Promise<InsertNoteCommitmentsEvent[]> = query(
-      this.contract,
-      this.contract.filters.InsertNoteCommitments(),
-      from,
-      to
-    );
-    const noteEventsProm: Promise<InsertNotesEvent[]> = query(
-      this.contract,
-      this.contract.filters.InsertNotes(),
-      from,
-      to
-    );
-
-    const [noteCommitmentEvents, noteEvents] = await Promise.all([
-      ncEventsProm,
-      noteEventsProm,
-    ]);
-
-    // extract leaves from each (note commitments are the leaves, full notes have to be hashed)
-    // combine them into a single list
-    // and sort them in the order in which they appeared on-chain
-
-    let leaves: OrderedLeaf[] = [];
-    for (const event of noteCommitmentEvents) {
-      const eventLeaves = event.args.commitments.map((l) => l.toBigInt());
-      const orderedLeaves = eventLeaves.map((leaf) => ({
-        leaf,
-        blockNumber: event.blockNumber,
-        txIdx: event.transactionIndex,
-        logIdx: event.logIndex,
-      }));
-      leaves.push(...orderedLeaves);
-    }
-
-    for (const event of noteEvents) {
-      for (const noteValues of event.args.notes) {
-        const owner = {
-          h1X: noteValues.ownerH1.toBigInt(),
-          h2X: noteValues.ownerH2.toBigInt(),
-          h1Y: 0n,
-          h2Y: 0n,
-        };
-
-        const note = {
-          owner,
-          nonce: noteValues.nonce.toBigInt(),
-          asset: noteValues.asset.toHexString(),
-          id: noteValues.id.toBigInt(),
-          value: noteValues.value.toBigInt(),
-        };
-
-        const leaf = NoteTrait.toCommitment(note);
-        leaves.push({
-          leaf,
-          blockNumber: event.blockNumber,
-          txIdx: event.transactionIndex,
-          logIdx: event.logIndex,
-        });
+    const insertions = await fetchInsertions(this.contract, from, to);
+    return insertions.map((insertion) => {
+      if (typeof insertion === "bigint") {
+        // it's a note commitment
+        return insertion;
+      } else {
+        // it's a note
+        return NoteTrait.toCommitment(insertion);
       }
-    }
-
-    leaves = leaves.sort(
-      (a, b) =>
-        a.blockNumber - b.blockNumber ||
-        a.txIdx - b.txIdx ||
-        a.logIdx - b.logIdx
-    );
-
-    // return only the leaves
-    return leaves.map(({ leaf }) => leaf);
+    });
   }
 }