+++ packages/sdk/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 
+- Add interfaces for subtree update prover
 - Remove unnessary normalization of circuit values
 - Add context methods for ensuring min asset balance and reading all balances
 - Proving no longer takes files as function args (moved to local-prover class instantiation)